requirements.txt
torch>=2.3.0
numpy>=1.24.0
scipy>=1.10.0
einops>=0.7.0
faiss-gpu>=1.7.0
biopython>=1.81
pyhmmer>=0.10.0
sentence-transformers>=2.2.0
pandas>=2.0.0
pyarrow>=12.0.0
tensorboard>=2.13.0
wandb>=0.15.0
pyyaml>=6.0
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0
